refactor(api): split trpc routers/structure by entity vs link
Extracts link/member-link/association procedures into
routers/structure/links.ts; barrel composes via spread.
Depends on the trpc-parity recursion fix in the previous commit.

Bean: api-x4d7
