+++ apps/api/src/trpc/routers/structure.ts
@@ -1,22 +1,13 @@
 import {
-  CreateStructureEntityAssociationBodySchema,
   CreateStructureEntityBodySchema,
-  CreateStructureEntityLinkBodySchema,
-  CreateStructureEntityMemberLinkBodySchema,
   CreateStructureEntityTypeBodySchema,
   UpdateStructureEntityBodySchema,
-  UpdateStructureEntityLinkBodySchema,
   UpdateStructureEntityTypeBodySchema,
   brandedIdQueryParam,
 } from "@pluralscape/validation";
 import { z } from "zod/v4";
 
-import {
-  createEntityAssociation,
-  deleteEntityAssociation,
-  getEntityHierarchy,
-  listEntityAssociations,
-} from "../../services/structure/association.js";
+import { getEntityHierarchy } from "../../services/structure/association.js";
 import { createStructureEntity } from "../../services/structure/entity-crud/create.js";
 import {
   archiveStructureEntity,
@@ -35,20 +26,11 @@ import { getEntityType } from "../../services/structure/entity-type/get.js";
 import { listEntityTypes } from "../../services/structure/entity-type/list.js";
 import { restoreEntityType } from "../../services/structure/entity-type/restore.js";
 import { updateEntityType } from "../../services/structure/entity-type/update.js";
-import {
-  createEntityLink,
-  deleteEntityLink,
-  listEntityLinks,
-  updateEntityLink,
-} from "../../services/structure/link.js";
-import {
-  createEntityMemberLink,
-  deleteEntityMemberLink,
-  listEntityMemberLinks,
-} from "../../services/structure/member-link.js";
 import { createTRPCCategoryRateLimiter } from "../middlewares/rate-limit.js";
 import { systemProcedure } from "../middlewares/system.js";
 import { router } from "../trpc.js";
+
+import { linkProcedures } from "./structure/links.js";
 
 const readLimiter = createTRPCCategoryRateLimiter("readDefault");
 const writeLimiter = createTRPCCategoryRateLimiter("write");
@@ -62,18 +44,6 @@ const EntityTypeIdSchema = z.object({
 
 const EntityIdSchema = z.object({
   entityId: brandedIdQueryParam("ste_"),
-});
-
-const LinkIdSchema = z.object({
-  linkId: brandedIdQueryParam("stel_"),
-});
-
-const MemberLinkIdSchema = z.object({
-  memberLinkId: brandedIdQueryParam("steml_"),
-});
-
-const AssociationIdSchema = z.object({
-  associationId: brandedIdQueryParam("stea_"),
 });
 
 export const structureRouter = router({
@@ -249,126 +219,5 @@ export const structureRouter = router({
       }),
   }),
 
-  // ── Entity Links ─────────────────────────────────────────────────
-
-  link: router({
-    create: systemProcedure
-      .use(writeLimiter)
-      .input(CreateStructureEntityLinkBodySchema)
-      .mutation(async ({ ctx, input }) => {
-        const audit = ctx.createAudit(ctx.auth);
-        return createEntityLink(ctx.db, ctx.systemId, input, ctx.auth, audit);
-      }),
-
-    list: systemProcedure
-      .use(readLimiter)
-      .input(
-        z.object({
-          cursor: z.string().nullish(),
-          limit: z.number().int().min(1).max(MAX_LIST_LIMIT).optional(),
-        }),
-      )
-      .query(async ({ ctx, input }) => {
-        return listEntityLinks(ctx.db, ctx.systemId, ctx.auth, {
-          cursor: input.cursor ?? undefined,
-          limit: input.limit,
-        });
-      }),
-
-    update: systemProcedure
-      .use(writeLimiter)
-      .input(LinkIdSchema.and(UpdateStructureEntityLinkBodySchema))
-      .mutation(async ({ ctx, input }) => {
-        const audit = ctx.createAudit(ctx.auth);
-        return updateEntityLink(
-          ctx.db,
-          ctx.systemId,
-          input.linkId,
-          { sortOrder: input.sortOrder },
-          ctx.auth,
-          audit,
-        );
-      }),
-
-    delete: systemProcedure
-      .use(writeLimiter)
-      .input(LinkIdSchema)
-      .mutation(async ({ ctx, input }) => {
-        const audit = ctx.createAudit(ctx.auth);
-        await deleteEntityLink(ctx.db, ctx.systemId, input.linkId, ctx.auth, audit);
-        return { success: true as const };
-      }),
-  }),
-
-  // ── Entity Member Links ──────────────────────────────────────────
-
-  memberLink: router({
-    create: systemProcedure
-      .use(writeLimiter)
-      .input(CreateStructureEntityMemberLinkBodySchema)
-      .mutation(async ({ ctx, input }) => {
-        const audit = ctx.createAudit(ctx.auth);
-        return createEntityMemberLink(ctx.db, ctx.systemId, input, ctx.auth, audit);
-      }),
-
-    list: systemProcedure
-      .use(readLimiter)
-      .input(
-        z.object({
-          cursor: z.string().nullish(),
-          limit: z.number().int().min(1).max(MAX_LIST_LIMIT).optional(),
-        }),
-      )
-      .query(async ({ ctx, input }) => {
-        return listEntityMemberLinks(ctx.db, ctx.systemId, ctx.auth, {
-          cursor: input.cursor ?? undefined,
-          limit: input.limit,
-        });
-      }),
-
-    delete: systemProcedure
-      .use(writeLimiter)
-      .input(MemberLinkIdSchema)
-      .mutation(async ({ ctx, input }) => {
-        const audit = ctx.createAudit(ctx.auth);
-        await deleteEntityMemberLink(ctx.db, ctx.systemId, input.memberLinkId, ctx.auth, audit);
-        return { success: true as const };
-      }),
-  }),
-
-  // ── Entity Associations ──────────────────────────────────────────
-
-  association: router({
-    create: systemProcedure
-      .use(writeLimiter)
-      .input(CreateStructureEntityAssociationBodySchema)
-      .mutation(async ({ ctx, input }) => {
-        const audit = ctx.createAudit(ctx.auth);
-        return createEntityAssociation(ctx.db, ctx.systemId, input, ctx.auth, audit);
-      }),
-
-    list: systemProcedure
-      .use(readLimiter)
-      .input(
-        z.object({
-          cursor: z.string().nullish(),
-          limit: z.number().int().min(1).max(MAX_LIST_LIMIT).optional(),
-        }),
-      )
-      .query(async ({ ctx, input }) => {
-        return listEntityAssociations(ctx.db, ctx.systemId, ctx.auth, {
-          cursor: input.cursor ?? undefined,
-          limit: input.limit,
-        });
-      }),
-
-    delete: systemProcedure
-      .use(writeLimiter)
-      .input(AssociationIdSchema)
-      .mutation(async ({ ctx, input }) => {
-        const audit = ctx.createAudit(ctx.auth);
-        await deleteEntityAssociation(ctx.db, ctx.systemId, input.associationId, ctx.auth, audit);
-        return { success: true as const };
-      }),
-  }),
+  ...linkProcedures,
 });
